alter priority to decmail in task and section

+++ task/models.py
@@ -4,7 +4,7 @@ from section.models import Section
 
 class Task(BaseModel):
 	user = models.ForeignKey('auth.User', on_delete=models.CASCADE)
-	priority = models.IntegerField()
+	priority = models.DecimalField(max_digits=20, decimal_places=10)
 	section = models.ForeignKey(Section, on_delete=models.CASCADE, related_name='tasks')
 	description = models.TextField()
 	name = models.CharField(max_length=255)
